backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.127.1",
    "pydantic>=2.9.0",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.9",
    "uvicorn[standard]>=0.40.0",
    "python-jose[cryptography]>=3.5.0",
]
